requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
distlib==0.3.9
dulwich==0.21.7
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jose==1.0.0
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
mysqlclient==2.1.1
packaging==24.2
passlib==1.7.4
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
PyMySQL==1.1.1
pyproject_hooks==1.2.0
pytest==8.3.4
pytest-mock==3.14.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
RapidFuzz==3.11.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.28.0
xattr==1.1.0
zipp==3.21.0